Move decodeAccount to utils module
