+++ src/components/AddAccountDialog.js
@@ -7,7 +7,7 @@ import TextField from '@material-ui/core/TextField';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import FormGroup from '@material-ui/core/FormGroup';
 import Switch from '@material-ui/core/Switch';
-import { decodeAccount } from './helpers/import';
+import { decodeAccount } from '../utils/utils';
 import DialogForm from './DialogForm';
 
 export default function AddAccountDialog({ open, onAdd, onClose }) {
