+++ src/utils/utils.ts
@@ -1,5 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
-import { Connection, PublicKey } from '@solana/web3.js';
+import { Account, Connection, PublicKey } from '@solana/web3.js';
+import * as bs58 from 'bs58';
 
 export async function sleep(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms));
@@ -84,4 +85,20 @@ export async function confirmTransaction(
     new Date().getTime() - startTime.getTime(),
   );
   return result.value;
+}
+
+/**
+ * Returns an account object when given the private key
+ *
+ */
+export const decodeAccount = (privateKey: string) => {
+  try {
+    return new Account(JSON.parse(privateKey));
+  } catch (_) {
+    try {
+      return new Account(bs58.decode(privateKey));
+    } catch (_) {
+      return undefined;
+    }
+  }
 }
